Also move `StarlightRouteData` type
