+++ packages/starlight/props.ts
@@ -1,2 +1,2 @@
-export type { StarlightRouteData as Props } from './utils/routing/data';
+export type { StarlightRouteData as Props } from './utils/routing/types';
 export type { StarlightPageProps } from './utils/starlight-page';
