+++ packages/starlight/utils/starlight-page.ts
@@ -5,14 +5,8 @@ import config from 'virtual:starlight/user-config';
 import { getCollectionPathFromRoot } from './collection';
 import { parseWithFriendlyErrors, parseAsyncWithFriendlyErrors } from './error-map';
 import { stripLeadingAndTrailingSlashes } from './path';
-import {
-	getSiteTitle,
-	getSiteTitleHref,
-	getToC,
-	type PageProps,
-	type StarlightRouteData,
-} from './routing/data';
-import type { StarlightDocsEntry } from './routing/types';
+import { getSiteTitle, getSiteTitleHref, getToC, type PageProps } from './routing/data';
+import type { StarlightDocsEntry, StarlightRouteData } from './routing/types';
 import { slugToLocaleData, urlToSlug } from './slugs';
 import { getPrevNextLinks, getSidebar, getSidebarFromConfig } from './navigation';
 import { docsSchema } from '../schema';
